// src/lib/models/authenticate-user-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * API Documentation
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

 /**
 * 
 *
 * @export
 * @interface AuthenticateUserDto
 */
export interface AuthenticateUserDto {

    /**
     * @type {string}
     * @memberof AuthenticateUserDto
     */
    projectId: string;

    /**
     * @type {string}
     * @memberof AuthenticateUserDto
     */
    login: string;

    /**
     * @type {string}
     * @memberof AuthenticateUserDto
     */
    secret: string;
}