// src/lib/models/authentication-resource-organization.ts
/* tslint:disable */
/* eslint-disable */
/**
 * API Documentation
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { Organization } from './organization';
 /**
 * 
 *
 * @export
 * @interface AuthenticationResourceOrganization
 */
export interface AuthenticationResourceOrganization {

    /**
     * @type {Organization}
     * @memberof AuthenticationResourceOrganization
     */
    data: Organization;

    /**
     * @type {string}
     * @memberof AuthenticationResourceOrganization
     */
    token: string;
}